+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
-### v5.18.0.0 (uncut)
+### v5.18.0.0
 - **http4k-connect-*** - Upgrade dependencies.
 - **http4k-connect-amazon-cognito*** - [Breaking] AWS Cognito: Add support for server side authentication (
   AdminInitiateAuth and AdminRespondToAuthChallenge). H/T @markth0mas